refactor(network): move stream service logic into createLatestEventStreamService util
