+++ packages/network/src/workers/SyncWorker.ts
@@ -1,5 +1,5 @@
 import { awaitStreamValue, DoWork, streamToDefinedComputed } from "@latticexyz/utils";
-import { expand, from, map, Observable, Subject, takeWhile } from "rxjs";
+import { Observable, Subject } from "rxjs";
 import { NetworkComponentUpdate, SyncWorkerConfig } from "../types";
 import { Components } from "@latticexyz/recs";
 import {
@@ -19,12 +19,10 @@ import {
   createDecode,
   createFetchWorldEventsInBlockRange,
   getSnapshotBlockNumber,
-  fetchSnapshot,
   fetchStateInBlockRange,
   fetchSnapshotChunked,
-  createStreamClient,
   createLatestEventStreamRPC,
-  convertStreamServiceMessageToEvent,
+  createLatestEventStreamService,
 } from "./syncUtils";
 import { createBlockNumberStream } from "../createBlockNumberStream";
 export type Output<Cm extends Components> = NetworkComponentUpdate<Cm>;
@@ -80,29 +78,14 @@ export class SyncWorker<Cm extends Components> implements DoWork<SyncWorkerConfi
     let passLiveEventsToOutput = false;
     const cacheStore = { current: createCacheStore() };
     const { blockNumber$ } = createBlockNumberStream(providers);
+    const latestEventStream = streamServiceUrl
+      ? createLatestEventStreamService(streamServiceUrl, worldContract.address)
+      : createLatestEventStreamRPC(blockNumber$, fetchWorldEvents);
 
-    const streamServiceClient = streamServiceUrl ? createStreamClient(streamServiceUrl) : undefined;
-    if (streamServiceClient) {
-      const stream = streamServiceClient.subscribeToStreamLatest({
-        worldAddress: worldContract.address,
-        blockNumber: true,
-        blockHash: true,
-        blockTimestamp: true,
-        transactionsConfirmed: true,
-        ecsEvents: true,
-      });
-
-      for await (const message of stream.responses) {
-        const event = convertStreamServiceMessageToEvent(message);
-        storeEvent(cacheStore.current, event);
-        if (passLiveEventsToOutput) this.output$.next(event as Output<Cm>);
-      }
-    } else {
-      createLatestEventStreamRPC(blockNumber$, fetchWorldEvents).subscribe((event) => {
-        storeEvent(cacheStore.current, event);
-        if (passLiveEventsToOutput) this.output$.next(event as Output<Cm>);
-      });
-    }
+    latestEventStream.subscribe((event) => {
+      storeEvent(cacheStore.current, event);
+      if (passLiveEventsToOutput) this.output$.next(event as Output<Cm>);
+    });
 
     const streamStartBlockNumber = awaitStreamValue(blockNumber$);
